% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{build.bin.sim.spec}
\alias{build.bin.sim.spec}
\title{A Cat Function}
\usage{
build.bin.sim.spec(bin.mass, target, bin.size = 1)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
