% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{build_bin_sim_spec}
\alias{build_bin_sim_spec}
\title{build_bin_sim_spec}
\usage{
build_bin_sim_spec(bin.mass, target, bin.size = 1)
}
\arguments{
\item{bin.mass}{Desciption}

\item{target}{Desciption}

\item{bin.size}{Desciption}
}
\description{
Internal file.
}
