% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{create.insilico.mixture.template}
\alias{create.insilico.mixture.template}
\title{A Cat Function}
\usage{
create.insilico.mixture.template(mono.info, mz.tol = 0.5)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
