% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{create_insilico_mixture_template}
\alias{create_insilico_mixture_template}
\title{create_insilico_mixture_template}
\usage{
create_insilico_mixture_template(mono.info, mz.tol = 0.5)
}
\arguments{
\item{mono.info}{An output of \code{\link{gather_summary}}.}

\item{mz.tol}{A m/z tolerance in Da. (Default: 0.5)}
}
\value{
A data frame which contains simulated m/z, log intensity, and normalized intensity values of peaks.
}
\description{
This function generates an intial template for simulated mass spectra.
}
\examples{
\dontrun{
create_insilico_mixture_template(mono.info)
}
}
