% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{gather.summary}
\alias{gather.summary}
\title{A Cat Function}
\usage{
gather.summary(x)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
