% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{gather_summary_file}
\alias{gather_summary_file}
\title{gather_summary_file}
\usage{
gather_summary_file(directory)
}
\arguments{
\item{directory}{A directory that contains summary files from \code{\link{summarize.monospectra}}.}
}
\value{
A list of combined summary and the corresponding species.
}
\description{
This function combines output files from \code{\link{summarize.monospectra}}.
}
\examples{
\dontrun{
gather.summary.file(directory='C:/')
}
}
