% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{install_check}
\alias{install_check}
\title{install_check}
\usage{
install_check()
}
\value{
Print MGMS2 install_check in R concole.
}
\description{
install_check function.
}
\examples{
install_check()
}
