% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{process.monospectra}
\alias{process.monospectra}
\title{A Cat Function}
\usage{
process.monospectra(file, mass.range = c(1000, 2200),
  halfWindowSize = 20, SNIP.iteration = 60)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
