% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{read.summary.file}
\alias{read.summary.file}
\title{A Cat Function}
\usage{
read.summary.file(files)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
