% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{simulate.ind.spec.single}
\alias{simulate.ind.spec.single}
\title{A Cat Function}
\usage{
\method{simulate}{ind.spec.single}(interest, mz.tol, species, strain)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
