% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{simulate_ind_spec_single}
\alias{simulate_ind_spec_single}
\title{simulate_ind_spec_single}
\usage{
simulate_ind_spec_single(interest, mz.tol, species, strain)
}
\arguments{
\item{interest}{Add description}

\item{mz.tol}{Add description}

\item{species}{Add description}

\item{strain}{Add description}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
