% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{summarize_monospectra}
\alias{summarize_monospectra}
\title{summarize_monospectra}
\usage{
summarize_monospectra(
  processed.obj,
  species,
  directory,
  minFrequency = 0.5,
  align.tolerance = 5e-04,
  snr = 3,
  halfWindowSize = 20,
  top.N = 50
)
}
\arguments{
\item{processed.obj}{Add description}

\item{species}{Add description}

\item{directory}{Add description}

\item{minFrequency}{Add description}

\item{align.tolerance}{Add description}

\item{snr}{Add description}

\item{halfWindowSize}{Add description}

\item{top.N}{Add description}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
