% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{summary.mono}
\alias{summary.mono}
\title{A Cat Function}
\usage{
\method{summary}{mono}(spectra.interest, minFrequency = 0.5,
  align.tolerance = 5e-04, snr = 3, halfWindowSize = 20,
  top.N = 50)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
