% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{summary_mono}
\alias{summary_mono}
\title{summary_mono}
\usage{
summary_mono(
  spectra.interest,
  minFrequency = 0.5,
  align.tolerance = 5e-04,
  snr = 3,
  halfWindowSize = 20,
  top.N = 50
)
}
\arguments{
\item{spectra.interest}{Description}

\item{minFrequency}{Description}

\item{align.tolerance}{Description}

\item{snr}{Description}

\item{halfWindowSize}{Description}

\item{top.N}{Description}
}
\description{
Internal file.
}
