% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{testfunction}
\alias{testfunction}
\title{testfunction}
\usage{
testfunction()
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
testfunction
}
\examples{
testfunction()
}
\keyword{cats}
