% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{create_insilico_mixture_template}
\alias{create_insilico_mixture_template}
\title{create_insilico_mixture_template}
\usage{
create_insilico_mixture_template(mono.info, mz.tol = 0.5)
}
\arguments{
\item{mono.info}{Add description}

\item{mz.tol}{Add description}
}
\description{
This function allows you to express your love of cats.
}
\examples{
create_insilico_mixture_template()
}
\keyword{cats}
