% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{filtermass}
\alias{filtermass}
\title{filtermass}
\usage{
filtermass(spectra, mass.range)
}
\arguments{
\item{spectra}{Spectra. A MALDIquant object. An output of \code{\link[MALDIquantForeign]{importMzXml}}.}

\item{mass.range}{Mass (m/z) range. For exmaple, c(1000,2200).}
}
\value{
The filtered spectra.
}
\description{
Internal function. This function removes peaks with their mass values (m/z values) outside a given mass range.
This function is used in \code{\link{process_monospectra}}.
}
