% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{filtermass}
\alias{filtermass}
\title{filtermass}
\usage{
filtermass(spectra, mass.range)
}
\arguments{
\item{spectra}{Add description}

\item{mass.range}{Add description}
}
\description{
This function allows you to express your love of cats.
}
\examples{
filtermass()
}
\keyword{cats}
