% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{filtermass}
\alias{filtermass}
\title{filtermass}
\usage{
filtermass(spectra, mass.range)
}
\arguments{
\item{spectra}{Description}

\item{mass.range}{Description}
}
\description{
Internal function.
}
