% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{preprocessMS}
\alias{preprocessMS}
\title{preprocessMS}
\usage{
preprocessMS(spectra, halfWindowSize = 20, SNIP.iteration = 60)
}
\arguments{
\item{spectra}{Add description}

\item{halfWindowSize}{Add description}

\item{SNIP.iteration}{Add description}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
