% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{preprocessMS}
\alias{preprocessMS}
\title{preprocessMS}
\usage{
preprocessMS(spectra, halfWindowSize = 20, SNIP.iteration = 60)
}
\arguments{
\item{spectra}{Spectra. A MALDIquant object. An output of either \code{\link[MALDIquantForeign]{importMzXml}} or \code{\link{filtermass}}.}

\item{halfWindowSize}{halfWindowSize The highest peaks in the given window (+/-halfWindowSize) will be recognized as peaks. (Default: 20). See \code{\link[MALDIquant]{detectPeaks}} for details.}

\item{SNIP.iteration}{SNIP.iteration An iteration used to remove the baseline of an spectrum. (Default: 60). See \code{\link[MALDIquant]{removeBaseline}} for details.}
}
\value{
The processed mass spectra. A list of MALDIquant MassSpectrum objects (S4 objects).
}
\description{
Internal function. This function preprocesses spectra by transforming/smoothing intensity, removing baseline, and calibrating intensities.
}
