% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{process_monospectra}
\alias{process_monospectra}
\title{process_monospectra}
\usage{
process_monospectra(
  file,
  mass.range = c(1000, 2200),
  halfWindowSize = 20,
  SNIP.iteration = 60
)
}
\arguments{
\item{file}{Add description}

\item{mass.range}{Add description}

\item{halfWindowSize}{Add description}

\item{SNIP.iteration}{Add description}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
