% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{process_monospectra}
\alias{process_monospectra}
\title{process_monospectra}
\usage{
process_monospectra(
  file,
  mass.range = c(1000, 2200),
  halfWindowSize = 20,
  SNIP.iteration = 60
)
}
\arguments{
\item{file}{A file name. This file is a tab-delimited file which contains the following columns: file names, strain.no, and strain. See below for details.}

\item{mass.range}{The m/z range that an user want to consider for the analysis. (Default: c(1000,2200)).}

\item{halfWindowSize}{A half window size used for the smoothing the intensity values. (Default: 20). See \code{\link[MALDIquant]{smoothIntensity}} for details.}

\item{SNIP.iteration}{An interaction used to remove the baseline of an spectrum. (Default: 60). See \code{\link[MALDIquant]{removeBaseline}} for details.}
}
\value{
A list of processed monobacterial spectra, and their strain numbers, a vector of unique strains, and strain names.
}
\description{
This function processes multiple mzXML files which are listed in the file that an user specifies.
}
\details{
This tab-delimited file contains a sample desription. 
file.name	strain.no	strain
Ab1.mzXML	1	EAS025
Ab2.mzXML	1	EAS025
Ab3.mzXML	1	EAS025
Ab4.mzXML	2	EAS028
Ab5.mzXML	2	EAS028
Ab6.mzXML	2	EAS028
Ab7.mzXML	3	SM1660
Ab8.mzXML	3	SM1660
Ab9.mzXML	3	SM1660
}
\examples{
\dontrun{
spectra.processed.Ab <- process_monospectra(file='Data/Acinetobacter baumannii - sen/list.txt', mass.range=c(1000,2200))
}
}
