% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions.R
\name{simulate_ind_spec_single}
\alias{simulate_ind_spec_single}
\title{simulate_ind_spec_single}
\usage{
simulate_ind_spec_single(interest, mz.tol, species, strain)
}
\arguments{
\item{interest}{Summary statistics of spectra.}

\item{mz.tol}{The tolerance of m/z. This is used to generate m/z values of peaks.}

\item{species}{Species.}

\item{strain}{Strain name.}
}
\value{
A data frame that contains m/z, (normalized) intensity values, missing rates of peaks, species name, and strain name.
}
\description{
Internal function. The function simulates m/z and intensity values using given summary statistics.
}
