% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{simulate_many_poly_spectra}
\alias{simulate_many_poly_spectra}
\title{simulate_many_poly_spectra}
\usage{
simulate_many_poly_spectra(mono.info, nsim = 10000,
  file = "MGMS2_insilico_spectra.pdf", mixture.ratio,
  mixture.missing.prob.peak = 0.05, noise.peak.ratio = 0.05,
  snr.basepeak = 500, noise.cv = 0.25, mz.range = c(1000, 2200),
  mz.tol = 0.5)
}
\arguments{
\item{love}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
testfunction()
}
\keyword{cats}
