% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpecFunctions_Final3.R
\name{summarize_monospectra}
\alias{summarize_monospectra}
\title{summarize_monospectra}
\usage{
summarize_monospectra(
  processed.obj,
  species,
  directory,
  minFrequency = 0.5,
  align.tolerance = 5e-04,
  snr = 3,
  halfWindowSize = 20,
  top.N = 50
)
}
\arguments{
\item{processed.obj}{A list from \code{\link{process.monospectra}} which contains top.N peaks information for each strain.}

\item{species}{Species name.}

\item{directory}{Directory.}

\item{minFrequency}{Percentage value. A minimum occurrence proportion required for building a reference peaks. All peaks with their occurence proportion less than minFrequency will be moved. (Default: 0.50). See \code{\link[MALDIquant]{filterPeaks}} and \code{\link[MALDIquant]{referencePeaks}} for details.}

\item{align.tolerance}{Mass tolerance. Must be multiplied by 10^-6 for ppm. (Default: 0.0005).}

\item{snr}{Signal-to-noise ratio. (Default: 3).}

\item{halfWindowSize}{The highest peaks in the given window (+/-halfWindowSize) will be recognized as peaks. (Default: 20). See \code{\link[MALDIquant]{detectPeaks}} for details.}

\item{top.N}{The top N peaks will be chosen for the analysis. An integer value. (Default: 50).}
}
\value{
A data frame that contains the peaks informations: m/z, mean log intensity, standard deviation of log intensity, missing rate of peaks. In addition, it also contains species and strain information.
}
\description{
This function summarize monomicrobial spectra and write the summary in the specified directory.
}
\examples{
\dontrun{
spectra.processed.Ab <- process_monospectra(file='Data/Acinetobacter baumannii - sen/list.txt', 
mass.range=c(1000,2200))
}
\dontrun{
spectra.mono.summary.Ab <- summarize_monospectra(processed.obj=spectra.processed.Ab, species='Ab',
directory='Mix_Final')
}
}
